requirements.txt
Django~=3.0.8
djangorestframework~=3.11.0
djangorestframework-simplejwt~=4.4.0
spacy~=2.3.2
Scrapy~=2.2.0
Twisted~=20.3.0
feedparser~=5.2.1
asgiref~=3.2.10
itemadapter~=0.1.0
pytextrank~=2.0.2
plac~=1.1.3
thinc~=7.4.1